Implemented Trail and Rigid Compound Dendulum
Implemented a trail of motion in the animation visualisation. Implemented rigid compound dendulum system equations of motion. Added (commented out) code to determine the time taken for the first flip of a variety of initial angles. The simulation now includes centre of mass, moment of inertia, and more damping constants (for the friction in the hinges and the air resistances on the centres of mass of each arm and the bob masses themselves).